requirements.txt
# Core
kivy==2.2.1
kivymd==1.1.1
python-dotenv==1.0.0

# Supabase
supabase==2.3.1
gotrue==1.4.1
realtime==1.0.2
postgrest==1.0.0

# QR Code
qrcode==7.4.2
python-barcode==0.15.1
Pillow==10.0.0
zbarcam==1.0.1
pyzbar==0.1.9
kivy-garden.zbarcam==1.0.0

# Database
sqlalchemy==2.0.20
alembic==1.12.1

# Testing
pytest==7.4.0
pytest-mock==3.11.1

# Development
black==23.7.0
flake8==6.1.0

# Platform specific
plyer==2.1.0

# Async
httpx==0.24.1

# Date/Time
python-dateutil==2.8.2
